requirements.txt
Babel==2.6.0
backpack==0.1
bcrypt==3.1.4
blinker==1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cleo==0.6.6
click==6.7
Faker==0.7.18
feedgen==0.7.0
Flask==0.12.2
Flask-Caching==1.3.3
Flask-Limiter==1.0.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-orator==0.2.0
Flask-WTF==0.14.2
idna==2.7
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
limits==1.3
lxml==4.2.1
markdown2==2.3.5
MarkupSafe==1.0
orator==0.9.7
passlib==1.7.1
pastel==0.1.0
pendulum==1.5.1
psycopg2==2.7.4
pyaml==16.12.2
pycparser==2.18
Pygments==2.2.0
pylev==1.3.0
pysolr==3.7.0
python-dateutil==2.7.3
python-dotenv==0.8.2
python-redis-lock==3.2.0
pytz==2018.4
pytzdata==2018.5
PyYAML==3.12
raven==6.9.0
redis==2.10.6
requests==2.19.1
rq==0.10.0
simplejson==3.15.0
six==1.11.0
timeago==1.0.8
tzlocal==1.5.1
unicode-slugify==0.1.3
Unidecode==1.0.22
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
